% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr_clust.R, R/sen_spe.R
\name{sen_spe}
\alias{sen_spe}
\title{Clustering output sensitivity + specificity}
\usage{
sen_spe(clustering_output = emr, cluster_color_set = NULL, plot_feature = NULL)
}
\arguments{
\item{clustering_output}{\code{emr_clust} output}

\item{cluster_color_set}{for ggplot color set}

\item{plot_feature}{a vector indicating the indexes(colnames) of plot features}
}
\value{
\code{plot}. Profile for the sum of sensitivity and specificity.

sen_spe_plot
}
\description{
Clustering output sensitivity + specificity

Title
}
\examples{
library(FactoMineR)
data(tea)

emr_cat<-emr_clust(tea[,1:13],
                   method =c("MCAk"),
                   kmoCutoff = 0.5,
                   nclus=c(4:5),
                   ndim=c(2:3),
                   group="Tea",
                   nstart=50)

sen_spe_out<-sen_spe(clustering_output = emr_cat )
sen_spe_out
sen_spe_out$level
in_parameters<-names(tea[,1:7])
sen_spe_out<-sen_spe(clustering_output = emr_cat, plot_feature = in_parameters)
sen_spe_out
}
