% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr_clust.R, R/wind_rose_plot.R
\name{wind_rose_plot}
\alias{wind_rose_plot}
\title{Clustering output Visualization using Wind Rose plot}
\usage{
wind_rose_plot(
  clustering_output,
  plot_feature = NULL,
  plot_value,
  proportion_score = 0.7,
  cluster_color_set = NULL
)
}
\arguments{
\item{clustering_output}{\code{emr_clust} output}

\item{plot_feature}{a vector indicating the indexes(colnames) of plot features}

\item{plot_value}{a value want to choose for plot}

\item{proportion_score}{remove exam items in all clusters are less than some percentage (default: 0.7)}

\item{cluster_color_set}{for ggplot color set}
}
\value{
\code{list}. 1) \code{overlapped}: Profile by cluster - overlapped. 2) \code{seperated}: Profile by cluster - separated

wind & ros plot
}
\description{
\code{emr_clust} can be used for doing feature selection + dimention reduction + clustering with data Visualization
}
\examples{
library(FactoMineR)
data(tea)

emr_cat<-emr_clust(tea[,1:13],
                   method =c("MCAk"),
                   kmoCutoff = 0.5,
                   nclus=c(4:5),
                   ndim=c(2:3),
                   group="Tea",
                   nstart=50)

pic<-wind_rose_plot(clustering_output = emr_cat, plot_value = "Not" )
pic$overlapped
pic$seperated

in_parameters<-names(tea[,1:7])
pic<-wind_rose_plot(clustering_output = emr_cat, plot_value = "Not" ,plot_feature = in_parameters)
pic$overlapped
pic$seperated
}
